pyproject.toml
[tool.poetry]
name = "chatbot-pisos-turisticos"
version = "0.1.0"
description = "Chatbot para pisos turísticos con FastAPI y GPT-4"
authors = ["Tu Nombre <tuemail@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"  # 🔹 Esto soluciona el problema con Supabase
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
redis = "^5.0.1"
asyncpg = "^0.29.0"
supabase = "^2.13.0"
numpy = "^2.2.3"
python-dotenv = "^1.0.1"
openai = "^1.63.2"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.162.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"